+++ app/src/main/java/com/zackhsi/kiva/activities/ProfileActivity.java
@@ -1,25 +1,40 @@
-package com.zackhsi.kiva;
+package com.zackhsi.kiva.activities;
 
-import android.content.Intent;
 import android.support.v7.app.ActionBarActivity;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
+import android.widget.ImageView;
+
+import com.squareup.picasso.Picasso;
+import com.zackhsi.kiva.R;
+import com.zackhsi.kiva.models.User;
 
 
-public class BrowseActivity extends ActionBarActivity {
+public class ProfileActivity extends ActionBarActivity {
+    private ImageView ivUser;
+    private User user;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_browse);
+        setContentView(R.layout.activity_profile);
+
+        this.user = getStubUser();
+
+        findViews();
+
+        Picasso.with(this).load(user.getImageUrl()).into(ivUser);
     }
 
+    private void findViews() {
+        this.ivUser = (ImageView) findViewById(R.id.ivUser);
+    }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_browse, menu);
+        getMenuInflater().inflate(R.menu.menu_profile, menu);
         return true;
     }
 
@@ -30,13 +45,12 @@ public class BrowseActivity extends ActionBarActivity {
         // as you specify a parent activity in AndroidManifest.xml.
         int id = item.getItemId();
 
-        //noinspection SimplifiableIfStatement
-        if (id == R.id.miProfile) {
-            Intent i = new Intent(this, ProfileActivity.class);
-            startActivity(i);
-            return true;
-        }
-
         return super.onOptionsItemSelected(item);
+    }
+
+    private User getStubUser() {
+        User user = new User();
+        user.setImageId(1841451);
+        return user;
     }
 }
