+++ app/src/main/res/layout/activity_profile.xml
@@ -1,9 +1,17 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.zackhsi.kiva.ProfileActivity">
+    tools:context="com.zackhsi.kiva.activities.ProfileActivity">
 
-    <TextView android:text="Profile" android:layout_width="wrap_content"
-        android:layout_height="wrap_content" />
-
+    <ImageView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/ivUser"
+        android:layout_alignParentTop="true"
+        android:layout_alignParentLeft="true"
+        android:layout_alignParentStart="true"
+        android:layout_alignParentRight="true"
+        android:layout_alignParentEnd="true"
+        android:minHeight="100dp"
+        />
 </RelativeLayout>
