+++ app/src/main/res/menu/menu_profile.xml
@@ -1,4 +1,4 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools" tools:context="com.zackhsi.kiva.ProfileActivity">
+    xmlns:tools="http://schemas.android.com/tools" tools:context="com.zackhsi.kiva.activities.ProfileActivity">
 </menu>
